+++ CHANGELOG.md
@@ -69,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Villagers tools and clothing
 - Cultists clothing
 - You can start the game by pressing "enter" from the character selection menu
+- Added server-side character saving
 
 ### Changed